requirements.txt
django-filter
datetime
Django
gunicorn
setuptools